Fixed the problem with cyclical links.
Now, the run_tests.sh script accepts symlink definition files, creates the symlinks when
tests are run and cleans them up when the --clean option is passed.

Created helpers.sh, which is a helper script that runs functions associated with the
symlinks. Required as a separate script because xargs can't run bash builtins.